refactor: move cursist button
